api/requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
black==20.8b1
celery==5.0.0
cfgv==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
distlib==0.3.1
Django==3.1.2
djangorestframework==3.12.1
filelock==3.0.12
identify==1.5.6
kombu==5.0.2
marshmallow==3.8.0
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.0
pre-commit==2.7.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.10.15
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
vine==5.0.0
virtualenv==20.0.35
wcwidth==0.2.5
